I have crudability on Blogs
I fixed the capitalization issue with the migrations by running nuke-
pave.
I did so by changing the schema and the migration for blogs to lowercase
letters and running db:nuke-pave. I then added a commented out slash in
delete to allow that script to be run. With those changes I successfully
was able to crud on blog and without proper authentication you can't
edit create or destroy the blogs. Next on the list is to create comments